🦀: isolate Rust encode/decode Code
