+++ .github/workflows/release.yml
@@ -3,7 +3,7 @@
 #
 #    maturin generate-ci github
 #
-name: CI
+name: Release
 
 on:
   push:
